Fix a bug introduced by change in ros::Time constructor.
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@7206 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ trajectory_rollout/src/governor_node.cpp
@@ -48,7 +48,7 @@ GovernorNode::GovernorNode(std::vector<std_msgs::Point2DFloat32> footprint_spec)
 {
   robot_vel_.setIdentity();
   robot_vel_.frame_id_ = "base_link";
-  robot_vel_.stamp_ = (ros::Time)0ULL;
+  robot_vel_.stamp_ = ros::Time(uint64_t(0ULL));
 
   //so we can draw the local path
   advertise<std_msgs::Polyline2D>("local_path", 10);
